requirements.txt
PyYAML==6.0
numpy==1.21.0
rasterio==1.2.6
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
matplotlib==3.4.3
opencv-python==4.5.3.20210927
geopandas==0.9.0
tqdm==4.62.3